+++ lang/en/messages.php
@@ -5,4 +5,5 @@ declare(strict_types=1);
 return [
     'close' => 'Close',
     'dismiss' => 'Dismiss',
+    'no_results' => 'No data found',
 ];